Refactor JWT configuration by extracting JwtIssuerConfig and updating imports
