+++ src/main/java/com/epam/ta/reportportal/core/configs/security/converters/AbstractJwtConverter.java
@@ -1,6 +1,6 @@
 package com.epam.ta.reportportal.core.configs.security.converters;
 
-import com.epam.ta.reportportal.core.configs.security.MultiIdentityProviderConfig.JwtIssuerConfig;
+import com.epam.ta.reportportal.core.configs.security.JwtIssuerConfig;
 import java.util.Collection;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.authentication.AbstractAuthenticationToken;
